refactor: move authentication-related usecases to a single folder
