+++ src/data/protocols/db/AddAccountRepository.d.ts
@@ -1,4 +1,4 @@
-import { AddAccountModel } from '../../../domain/usecases/AddAccount'
+import { AddAccountModel } from '../../../domain/usecases/authentication/AddAccount'
 import AccountModel from '../../../domain/models/Account'
 
 export default interface AddAccountRepository {
