+++ src/domain/usecases/authentication/AuthenticateCredentials.d.ts
