+++ src/domain/usecases/authentication/AuthenticateUser.d.ts
