+++ src/presentation/controllers/authentication/login/LoginController.spec.ts
@@ -1,8 +1,8 @@
 import Controller from '../../../protocols/Controller'
 import LoginController from './LoginController'
 import HttpHelper from '../../../helpers/http/HttpHelper'
-import AuthenticateUser from '../../../../domain/usecases/AuthenticateUser'
-import AuthenticateCredentials from '../../../../domain/usecases/AuthenticateCredentials'
+import AuthenticateUser from '../../../../domain/usecases/authentication/AuthenticateUser'
+import AuthenticateCredentials from '../../../../domain/usecases/authentication/AuthenticateCredentials'
 import HttpRequest from '../../../protocols/HttpRequest'
 import ServerError from '../../../errors/ServerError'
 import Validation from '../../../protocols/Validation'
