+++ src/presentation/controllers/authentication/login/LoginController.ts
@@ -2,7 +2,7 @@ import Controller from '../../../protocols/Controller'
 import HttpRequest from '../../../protocols/HttpRequest'
 import HttpResponse from '../../../protocols/HttpResponse'
 import HttpHelper from '../../../helpers/http/HttpHelper'
-import AuthenticateUser from '../../../../domain/usecases/AuthenticateUser'
+import AuthenticateUser from '../../../../domain/usecases/authentication/AuthenticateUser'
 import Validation from '../../../protocols/Validation'
 
 export default class LoginController implements Controller {
