+++ src/presentation/controllers/authentication/signup/SignUpController.spec.ts
@@ -1,11 +1,11 @@
 import SignUpController from './SignUpController'
 import ServerError from '../../../errors/ServerError'
-import { AddAccount, AddAccountModel } from '../../../../domain/usecases/AddAccount'
+import { AddAccount, AddAccountModel } from '../../../../domain/usecases/authentication/AddAccount'
 import AccountModel from '../../../../domain/models/Account'
 import Validation from '../../../protocols/Validation'
 import HttpRequest from '../../../protocols/HttpRequest'
-import AuthenticateUser from '../../../../domain/usecases/AuthenticateUser'
-import AuthenticateCredentials from '../../../../domain/usecases/AuthenticateCredentials'
+import AuthenticateUser from '../../../../domain/usecases/authentication/AuthenticateUser'
+import AuthenticateCredentials from '../../../../domain/usecases/authentication/AuthenticateCredentials'
 import HttpHelper from '../../../helpers/http/HttpHelper'
 import EmailAlreadyInUseError from '../../../errors/EmailAlreadyInUseError'
